requirements.txt
PyDictionary~=2.0.1

Pillow~=9.4.0
customtkinter~=5.1.2